Support CMSIS and rename conflicting header file
Fixes #361
